Add S3 offload config to `wp-config.php`
This makes it possible to remove the configuration stored in the
database. This makes it safer to copy the database to your local
environment. It will not start to talk to production buckets.

This configuration supports AWS assume roles, that the preferred way to
authenticate AWS API calls from an EC2 instance.